requirements.txt
flask
python-dotenv
openai
picamera2
Pillow